Operators now live in the environment
** Interface changes

- Operators are now stored in the environment.
- Added Operator, makeOperator and makeBareOperator functionality
	This allows users to write operators that the parser can recognize during eval
- Expanded the number of default operators
- Added a macro substition system. This adds syntactic sugar to the operators

** Internal changes

- Added tokenyze.py [https://github.com/vigilantesculpting/tokenyze] to do the line tokenization, replacing shlex.

** Overview

The major change in this update is how operators are stored directly in the environment.
This should allow namespacing of operators in a suitable environment class, and just makes
things a bit tidier.

Operators have to be given proper names now, in order to add syntactic sugar a macro system
allows shortcuts for operator expansion.

For example, the squash operator is now a standalone operators, so instead of writing
	"2 2 2 2 2 +*"
we can now write
	"2 2 2 2 2 add squash"
Since 'sum' is a macro for 'add squash', we can instead simply write
	"2 2 2 2 2 sum"

Other syntactic sugar:

	"'users' 'jack' 'Desktop' //+" -> "users/jack/Desktop"
	"8 range 3 {}" -> [[0, 1, 2], [3, 4, 5], [6, 7]]
	"('jack' 'suzy' 'mary') 1 None []" -> ["suzy", mary"]