Refactor operators
** Interface changes

- MakeOperator is a factory function that adapts functions (especially stuff from
	the operator module) as Pretzyl.operators
- MakeOperator takes a bunch of arguments that make it possible to manipulate
	the pretzyl instance to achieve packing, unpacking, etc.
- The squash operator has been removed, it was completely broken (see below)

** Internal changes

- The makeoperator() method has been broken into two methods: getoperator() and runoperator().
This makes profiling of the two actions easier.
getoperator has been simplified so that detecting a valid operator is now faster.

** The squash operator modifier problem

The squash operator modifier did not work as intended.
The idea was that the squash would modify the operator that preceded it.
However, by the time the squash got called, the operator would have already had to fire at least
once.

So the following example would not work at all:
	2 add squash
since the add would have failed (stack underflow) before the squash got a chance to even run.

The solution is a new operator, called 'repeat'.
It takes a single string parameter which it turns into an operator, by looking up the string in
the environment, then executing it until it raises a stackunderflow:
	2 'add' repeat

This means the operator-being-modified never runs until repeat invokes it.